Modified "AsyncTaskFileReceiver" to show a progress dialog while receiving the audio file.

+++ app/src/main/java/org/marceloleite/projetoanna/utils/progressmonitor/ProgressReport.java
@@ -8,14 +8,14 @@ public class ProgressReport {
 
     private String message;
 
-    private double percentageConcluded;
+    private int percentageConcluded;
 
-    public ProgressReport(String message, double percentageConcluded) {
+    public ProgressReport(String message, int percentageConcluded) {
         this.percentageConcluded = percentageConcluded;
         this.message = message;
     }
 
-    public double getPercentageConcluded() {
+    public int getPercentageConcluded() {
         return percentageConcluded;
     }
